tests: Split test app logic into separate files
